Initial Commit
I have put a lot of work into setting up a fan site for Meridian 59,
however the feedback from the community was just not there. Putting this
project on hold so I can work on some smaller projects, but decided to
add this to GitHub incase someone else wanted to take a look.